split poll_pending_deposits() tests and create_stellar_deposit() tests
